# Speed up compile time evaluation of ImpliedDoLoop