The rest of the slowness was due to, 

https://github.com/lfortran/lfortran/blob/9465ce08a8263f65fbbdccbd16a43b031ff77606/src/lfortran/semantics/ast_common_visitor.h#L269

https://github.com/lfortran/lfortran/blob/9465ce08a8263f65fbbdccbd16a43b031ff77606/src/lfortran/semantics/ast_common_visitor.h#L243-L247

It queries a map. So I optimised the map to the most. We can't comment out the above line out (we need it). So this is the best that I could get.

```fortran
 program compile_time
implicit none
integer, parameter :: N = 10000000
integer, parameter :: sp = kind(1.0)
integer, parameter :: dp = kind(1.d0)
real(dp), parameter :: pi = 2*asin(1._dp)
real(dp), parameter :: a = 0, b = pi
real(dp), parameter :: dx = (b-a)/N
integer :: i
real(dp), parameter :: X(*) = [(sin(a+(b-a)*i/N), i = 1, N)]
real(dp), parameter :: S = sum(X)*dx
logical, parameter :: l = S < 2
real(kind=merge(sp, dp, l)) :: y
if (kind(y) == sp) then
print *, "true"
else
print *, "false"
end if
print *, S
end program
```

**Before**

`lfortran ../debug.f90 -o a.out  26.30s user 0.64s system 99% cpu 26.997 total`

**After**

`lfortran ../debug.f90 -o a.out  16.52s user 0.64s system 99% cpu 17.167 total`

Please report your results on the PR.


Fixes #5890.