@Pranavchiku 2025-01-03
Before

```console
% time lfortran a.f90
true
1.99999999999980860e+00
lfortran a.f90  27.05s user 1.16s system 99% cpu 28.418 total
```

After

```console
% time lfortran a.f90 
true
1.99999999999980860e+00
lfortran a.f90  17.38s user 0.86s system 99% cpu 18.406 total
```