@certik 2025-01-03
Great job, this fixes it for me: https://github.com/lfortran/lfortran/issues/5890#issuecomment-2569617529.